refactor: Removed test_utils
