+++ elasticai/creator/base_modules/torch_math_operations_test.py
@@ -1,6 +1,6 @@
 import torch
 
-from elasticai.creator.test_utils.tensor_test_case import TensorTestCase
+from tests.tensor_test_case import TensorTestCase
 
 from .torch_math_operations import TorchMathOperations
