+++ tests/tensor_test_case.py
